+++ src/test/kotlin/th/ac/kmitl/it/soa/group8/einvoice/models/SellerTest.kt
@@ -2,16 +2,16 @@ package th.ac.kmitl.it.soa.group8.einvoice.models
 
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
-import th.ac.kmitl.it.soa.group8.einvoice.model.Seller
+import th.ac.kmitl.it.soa.group8.e_tax_invoice.model.Seller
 
 class SellerTest {
     @Test
     fun sellerTest(){
-        val seller = Seller("3456","1234","Peter","5678",
-                        "Bob","customerService","abc@gmail.co.th",
-                        "022222222","A building","123 b road",
-                        "234 bangkok","rama 4","Prakanong","Prakanong",
-                        "234","112","890")
+        val seller = Seller("3456", "1234", "Peter", "5678",
+                "Bob", "customerService", "abc@gmail.co.th",
+                "022222222", "A building", "123 b road",
+                "234 bangkok", "rama 4", "Prakanong", "Prakanong",
+                "234", "112", "890")
         assertEquals("3456",seller.id)
         assertEquals("1234",seller.globalID)
         assertEquals("Peter",seller.name)
