+++ src/test/kotlin/th/ac/kmitl/it/soa/group8/einvoice/models/TradeContactModelTest.kt
@@ -2,7 +2,7 @@ package th.ac.kmitl.it.soa.group8.einvoice.models
 
 import com.thoughtworks.selenium.SeleneseTestBase.assertEquals
 import org.junit.Test
-import th.ac.kmitl.it.soa.group8.einvoice.model.TradeContactModel
+import th.ac.kmitl.it.soa.group8.e_tax_invoice.model.TradeContactModel
 
 class TradeContactModelTest {
     val personName: String = "Montita"
