+++ src/test/kotlin/th/ac/kmitl/it/soa/group8/einvoice/models/TypeCodeTest.kt
@@ -2,7 +2,7 @@ package th.ac.kmitl.it.soa.group8.einvoice.models
 
 import com.thoughtworks.selenium.SeleneseTestBase.assertEquals
 import org.junit.Test
-import th.ac.kmitl.it.soa.group8.einvoice.model.TypeCode
+import th.ac.kmitl.it.soa.group8.e_tax_invoice.model.TypeCode
 
 class TypeCodeTest {
